Refactor script fns into pallet.script.lib and make script fns use unquote
In order to remove the possibility of accidentally using a script function,
they must now be invoked using unquote. The use of unquote is a reminder that
clojure namespace resolution is in effect:
    (stevedore/script (~lib/rm file))

The script functions have been factored out of pallet.resource.* into
pallet.script.lib in order to improve the seperation of the target script
generation.