Add source form location in script comments
When generating script, insert the file and line of the source forms that
generate the script.  The file and line are inserted as comments on the line
preceding the generated code.  This is a convenient format that allows script
lines to be easily used in continuation lines, when chaining commands together
for instance.

Introduces pallet.stevedore/fragment, which is similar to p.s/script, except
that it disables source comments and ensures pipelines are kept on a single
line.  This is required in some expansion contexts.

The pallet.stevedore/with-source-line-comments macro can also be used to control
the output of source comments when using p.stevedore/script.